fix(wpf): store GlobalAssetClass as enum in tree metadata so filter works (#99)
The asset class ComboBox filter in the WPF navigation panel was silently
excluding every asset regardless of selection. NavigationMapper stored the
metadata value as (int)asset.Class, but FilterTreeNode checks
'value is GlobalAssetClass' — a type pattern that always fails on int.

Removed the cast so the value is stored as GlobalAssetClass. Added
NavigationMapperTests to guard against regressions.

+++ Financial.Application/Services/NavigationMapper.cs
@@ -66,7 +66,7 @@ internal static class NavigationMapper
                 ["ISIN"] = asset.ISIN,
                 ["Country"] = asset.Country,
                 ["LocalTypeCode"] = asset.LocalTypeCode,
-                ["GlobalAssetClass"] = (int)asset.Class,
+                ["GlobalAssetClass"] = asset.Class,
                 ["Quantity"] = asset.Quantity,
                 ["AveragePrice"] = asset.AveragePrice,
                 ["IsActive"] = asset.IsActive,
